fix: Use credential definition id as restriction for request templates
