+++ packages/legacy/core/__tests__/contexts/configuration.ts
@@ -2,7 +2,7 @@ import { ConfigurationContext } from '../../App'
 import Record from '../../App/components/record/Record'
 import { useNotifications } from '../../App/hooks/notifications'
 import { OCABundleResolver } from '../../App/types/oca'
-import { defaultProofRequestTemplates } from '../../verifier/constants'
+import { defaultProofRequestTemplates } from '../../verifier'
 
 const configurationContext: ConfigurationContext = {
   pages: () => [],
